+++ src/dashboardWebView/hooks/useThemeColors.tsx
@@ -7,11 +7,8 @@ export default function useThemeColors() {
   const { experimental } = useSettingsContext();
 
   const getColors = useCallback((defaultColors: string, themeColors: string) => {
-    if (experimental) {
-      return themeColors;
-    }
-
-    return defaultColors;
+    // The feature is now enabled by default
+    return themeColors;
   }, [experimental]);
 
   return {