+++ src/panelWebView/components/Fields/TextField.tsx
@@ -1,17 +1,21 @@
-import { PencilIcon } from '@heroicons/react/outline';
+import { PencilIcon, SparklesIcon } from '@heroicons/react/outline';
 import * as React from 'react';
 import { useCallback, useEffect, useMemo } from 'react';
 import { useRecoilState } from 'recoil';
-import { BaseFieldProps } from '../../../models';
+import { BaseFieldProps, PanelSettings } from '../../../models';
 import { RequiredFieldsAtom } from '../../state';
 import { FieldTitle } from './FieldTitle';
 import { FieldMessage } from './FieldMessage';
+import { messageHandler } from '@estruyf/vscode/dist/client';
+import { CommandToCode } from '../../CommandToCode';
 
 export interface ITextFieldProps extends BaseFieldProps<string> {
   singleLine: boolean | undefined;
   wysiwyg: boolean | undefined;
   limit: number | undefined;
   rows?: number;
+  name: string;
+  settings: PanelSettings;
   onChange: (txtValue: string) => void;
 }
 
@@ -25,11 +29,14 @@ export const TextField: React.FunctionComponent<ITextFieldProps> = ({
   description,
   value,
   rows,
+  name,
+  settings,
   onChange,
   required
 }: React.PropsWithChildren<ITextFieldProps>) => {
   const [, setRequiredFields] = useRecoilState(RequiredFieldsAtom);
   const [text, setText] = React.useState<string | null>(value);
+  const [loading, setLoading] = React.useState<boolean>(false);
 
   const onTextChange = (txtValue: string) => {
     setText(txtValue);
@@ -75,6 +82,37 @@ export const TextField: React.FunctionComponent<ITextFieldProps> = ({
     }
   }, [showRequiredState, isValid]);
 
+  const suggestDescription = () => {
+    setLoading(true);
+    messageHandler.request<string>(CommandToCode.aiSuggestDescription).then((suggestion) => {
+      setLoading(false);
+
+      if (suggestion) {
+        setText(suggestion);
+        onChange(suggestion);
+      }
+    }).catch(() => {
+      setLoading(false);
+    });
+  };
+
+  const actionElement = useMemo(() => {
+    if (!settings?.aiEnabled || settings.seo.descriptionField !== name) {
+      return;
+    }
+
+    return (
+      <button
+        className='metadata_field__title__action'
+        title={`Use Front Matter AI to suggest ${label?.toLowerCase()}`}
+        type='button'
+        onClick={() => suggestDescription()}
+        disabled={loading}>
+        <SparklesIcon />
+      </button>
+    );
+  }, [settings?.aiEnabled, name]);
+
   useEffect(() => {
     if (text !== value) {
       setText(value);
@@ -83,7 +121,15 @@ export const TextField: React.FunctionComponent<ITextFieldProps> = ({
 
   return (
     <div className={`metadata_field`}>
-      <FieldTitle label={label} icon={<PencilIcon />} required={required} />
+      {
+        loading && (
+          <div className='metadata_field__loading'>
+            Generating suggestion...
+          </div>
+        )
+      }
+
+      <FieldTitle label={label} actionElement={actionElement} icon={<PencilIcon />} required={required} />
 
       {wysiwyg ? (
         <React.Suspense fallback={<div>Loading field</div>}>